Update LibrarianTradeFinder to use new apis for dynamic enchantment registry
In Minecraft 1.21 the enchantments has been made configurable in stead
of being hard coded. This means that the former registry of enchantments
does not exist anymore.
We need to use the generic class RegistryEntry as a holder for Enchantments
and use RegistryEntry::getIdAsString if we want to compare instances.

Some render apis has also been changed. Like the BufferBuilder. Somehow
the Tessalator is not actually in use so I just commented it out. The gui
seems to work just fine.